Split Android AI runtime folders
