+++ apps/android/feature/ai/src/main/java/com/flashcardsopensourceapp/feature/ai/runtime/AiChatRuntime.kt
@@ -15,6 +15,26 @@ import com.flashcardsopensourceapp.data.local.repository.AutoSyncEventRepository
 import com.flashcardsopensourceapp.feature.ai.AiCardHandoffResult
 import com.flashcardsopensourceapp.feature.ai.AiEntryPrefill
 import com.flashcardsopensourceapp.feature.ai.aiEntryPrefillPrompt
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiAccessContext
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiChatRuntimeState
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiComposerPhase
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiConversationBootstrapState
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.canEditAiDraft
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.canEditAiDraftText
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.canManageAiDraftAttachments
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.canPrepareAiDraftInComposerPhase
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.shouldPrepareGuestAccess
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatBootstrapCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatDictationCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatLiveStreamCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatRuntimeLifecycleCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatSendCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatSessionCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.errors.AiAlertState
+import com.flashcardsopensourceapp.feature.ai.runtime.observability.AiChatBreadcrumb
+import com.flashcardsopensourceapp.feature.ai.runtime.observability.AiChatWarning
+import com.flashcardsopensourceapp.feature.ai.runtime.observability.recordAiChatBreadcrumb
+import com.flashcardsopensourceapp.feature.ai.runtime.observability.recordAiChatWarning
 import com.flashcardsopensourceapp.feature.ai.strings.AiTextProvider
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
