+++ apps/android/feature/ai/src/test/java/com/flashcardsopensourceapp/feature/ai/runtime/AiBootstrapErrorPresentationTest.kt
@@ -2,6 +2,8 @@ package com.flashcardsopensourceapp.feature.ai.runtime
 
 import com.flashcardsopensourceapp.data.local.ai.AiChatRemoteException
 import com.flashcardsopensourceapp.data.local.model.makeOfficialCloudServiceConfiguration
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatBootstrapBlockedException
+import com.flashcardsopensourceapp.feature.ai.runtime.observability.makeAiBootstrapErrorPresentation
 import com.flashcardsopensourceapp.feature.ai.strings.testAiTextProvider
 import java.io.IOException
 import org.junit.Assert.assertEquals
