+++ apps/android/feature/ai/src/test/java/com/flashcardsopensourceapp/feature/ai/runtime/AiChatDictationCoordinatorTest.kt
@@ -1,5 +1,10 @@
 package com.flashcardsopensourceapp.feature.ai.runtime
 
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiChatRuntimeState
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiComposerPhase
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.makeAiDraftState
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatDictationCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatSessionCoordinator
 import com.flashcardsopensourceapp.data.local.model.AiChatDictationState
 import com.flashcardsopensourceapp.data.local.model.AiChatTranscriptionResult
 import com.flashcardsopensourceapp.data.local.model.defaultAiChatServerConfig
