+++ apps/android/feature/ai/src/test/java/com/flashcardsopensourceapp/feature/ai/runtime/AiChatRuntimeBootstrapProvisioningTest.kt
@@ -8,6 +8,12 @@ import com.flashcardsopensourceapp.data.local.model.AiChatConversation
 import com.flashcardsopensourceapp.data.local.model.AiChatDraftState
 import com.flashcardsopensourceapp.data.local.model.CloudAccountState
 import com.flashcardsopensourceapp.data.local.model.makeDefaultAiChatPersistedState
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.AiConversationBootstrapState
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.makeAssistantStatusMessage
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.makeAiDraftState
+import com.flashcardsopensourceapp.feature.ai.runtime.conversation.makeUserMessage
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatBootstrapCoordinator
+import com.flashcardsopensourceapp.feature.ai.runtime.coordinators.AiChatRuntimeLifecycleCoordinator
 import java.io.IOException
 import java.net.MalformedURLException
 import java.net.SocketException
